🧹 isolate LoginField from ResultCard Edit view
